requirements.txt
Django==5.1.6
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.1

# Image support
Pillow>=9.0.0

# Form customization
django-widget-tweaks

# Authentication
django-allauth
social-auth-app-django

python-dotenv==0.21.0
python-dateutil
djangorestframework
djangorestframework-simplejwt


